# cmd/api/docs/swagger.yaml
definitions:
  models.AddNewOffer:
    properties:
      clientId:
        type: integer
      country:
        type: string
      flag:
        type: integer
      image:
        type: string
      imageHeight:
        type: integer
      imageWidth:
        type: integer
      offerHome:
        type: integer
      pageCount:
        type: integer
      position:
        type: integer
      showFrom:
        type: string
      storeURL:
        type: string
      storeURLTitle:
        type: string
      textLocale:
        type: string
      validFrom:
        type: string
      validTo:
        type: string
      validityTextLocale:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
  /add-offer:
    post:
      consumes:
      - application/json
      description: Add a new offer by passing in the whole offer details including
        offer validity date
      parameters:
      - description: Add a new offer
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.AddNewOffer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add A new Offer for an existing Country
      tags:
      - OFFERS
  /get-offer/{country}:
    get:
      consumes:
      - application/json
      description: Get Latest Offer from different Countries by passing country short
        form
      parameters:
      - description: country short form
        in: path
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Offer Based On Country
      tags:
      - OFFERS
swagger: "2.0"